+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 * Add error code `TaxAmountIsRequiredForAibSwedish` for attribute `tax_amount` in `transaction` key for AIB:Domestic transactions in Sweden
+* Add `exchangeRateQuoteId` to `TransactionGateway`
+* Add `ExchangeRateQuoteIdIsTooLong` to `validation_error_codes`
 * Add the following fields to `AndroidPayCard` and `ApplePayCard`:
   * `commercial`
   * `debit`